Add files via upload
Going through the new .json file I had noticed that the values for "DocumentType" would have spaces placed randomly before and after the comma. This resulted in my initial solution of splitting the string with the separator ", " not valid. 

In order to resolve this, my initial plan was to loop through the string and store a string into a variable until it encountered a space. When it would encounter the space, it would check the next value if it was a comma or space and if it wasn't it would continue with the loop and if it was a comma or space it would add the variable into a list and then null the variable. Then it would loop through the string until it encountered the next non comma/space and repeat until the end of the string. This seemed very brute force so i wanted to find a simpler method.

I then though of using the strip() function. Using strip() I would be able to loop through a string and whenever a split() occurred, it would strip the leading and trailing whitespaces of each value in the new list created by the split().

As for the second task, my initial solution still worked with the updated .json file. 
What I did here would be to loop through the objects in reverse order and when ever the key ['surveyType'] had the value of 'first' it would remove the object it was located in. The reason why we're looping it in reverse order is because when we're removing object while in a loop, it would remove the object but the position in the loop would not be updated. Meaning every time an object was removed, it would skip the next object. Going in reverse order fixes this issue.